Handling for RMQ connection errors
This makes the heartbeats less often; however the real issue is the `consumer_timeout` value settable in `/etc/rabbitmq/rabbitmq.conf`. It defaults to 1,800,000 ms (30 minutes) which can easily be shorter than the pipeline runtime for a group of 8 inputs on 4 workers writing to S3. It should then be configured by the user and RMQ container restarted if necessary.